Add explicit return if getSmallImagePreview fails

+++ lib/private/Preview/Generator.php
@@ -147,6 +147,10 @@ class Generator {
 			$crop = $specifications[0]['crop'] ?? false;
 			$preview = $this->getSmallImagePreview($previewFolder, $file, $mimeType, $previewVersion, $crop);
 
+			if ($preview === null) {
+				throw new NotFoundException('No image preview provided');
+			}
+
 			if ($preview->getSize() === 0) {
 				$preview->delete();
 				throw new NotFoundException('Cached preview size 0, invalid!');
@@ -226,8 +230,10 @@ class Generator {
 	/**
 	 * Generate a small image straight away without generating a max preview first
 	 * Preview generated is 256x256
+	 *
+	 * @throws NotFoundException
 	 */
-	private function getSmallImagePreview(ISimpleFolder $previewFolder, File $file, string $mimeType, string $prefix, bool $crop) {
+	private function getSmallImagePreview(ISimpleFolder $previewFolder, File $file, string $mimeType, string $prefix, bool $crop): ?ISimpleFile {
 		$nodes = $previewFolder->getDirectoryListing();
 
 		foreach ($nodes as $node) {
@@ -284,6 +290,8 @@ class Generator {
 				return $file;
 			}
 		}
+
+		return null;
 	}
 
 	/**
